Added a cast to avoid a warning from visual studio 2013

+++ dlib/graph_utils/edge_list_graphs.h
@@ -560,7 +560,7 @@ namespace dlib
             }
         }
         if (neighbors.size() != 0)
-            neighbors[cur_node] = std::make_pair(start_idx, edges.size());
+            neighbors[cur_node] = std::make_pair(start_idx, (unsigned long)edges.size());
     }
 
 // ----------------------------------------------------------------------------------------
